feat: extract BaseLayout
