+++ frontend/src/App.vue
@@ -9,7 +9,7 @@
 import { IonApp, IonRouterOutlet } from "@ionic/vue"
 import { defineComponent } from "vue"
 
-import Menu from "@/components/header/Menu.vue"
+import Menu from "@/components/Menu.vue"
 
 export default defineComponent({
 	name: "App",
