+++ frontend/src/components/QuickLinks.vue
